[FLINK-7968] [core] Move DataOutputSerializer and DataInputDeserializer to 'flink-core'
These core flink utils are independent of any other runtime classes and
are also used both in flink-runtime and in flink-queryable-state (which duplicated
the code).
