+++ flink-core/src/test/java/org/apache/flink/testutils/serialization/types/BooleanType.java
@@ -17,7 +17,7 @@
  */
 
 
-package org.apache.flink.runtime.io.network.api.serialization.types;
+package org.apache.flink.testutils.serialization.types;
 
 import java.io.IOException;
 import java.util.Random;
@@ -25,51 +25,50 @@ import java.util.Random;
 import org.apache.flink.core.memory.DataInputView;
 import org.apache.flink.core.memory.DataOutputView;
 
-public class CharType implements SerializationTestType {
+public class BooleanType implements SerializationTestType {
 
-	private char value;
+	private boolean value;
 
-	public CharType() {
-		this.value = 0;
+	public BooleanType() {
+		this.value = false;
 	}
 
-	private CharType(char value) {
+	private BooleanType(boolean value) {
 		this.value = value;
 	}
 
 	@Override
-	public CharType getRandom(Random rnd) {
-		return new CharType((char) rnd.nextInt(10000));
+	public BooleanType getRandom(Random rnd) {
+		return new BooleanType(rnd.nextBoolean());
 	}
 
 	@Override
 	public int length() {
-		return 2;
+		return 1;
 	}
 
 	@Override
 	public void write(DataOutputView out) throws IOException {
-		out.writeChar(this.value);
+		out.writeBoolean(this.value);
 	}
 
 	@Override
 	public void read(DataInputView in) throws IOException {
-		this.value = in.readChar();
+		this.value = in.readBoolean();
 	}
 
 	@Override
 	public int hashCode() {
-		return this.value;
+		return this.value ? 1 : 0;
 	}
 
 	@Override
 	public boolean equals(Object obj) {
-		if (obj instanceof CharType) {
-			CharType other = (CharType) obj;
+		if (obj instanceof BooleanType) {
+			BooleanType other = (BooleanType) obj;
 			return this.value == other.value;
 		} else {
 			return false;
 		}
 	}
 }
-
