+++ flink-core/src/test/java/org/apache/flink/testutils/serialization/types/ByteArrayType.java
@@ -17,56 +17,65 @@
  */
 
 
-package org.apache.flink.runtime.io.network.api.serialization.types;
+package org.apache.flink.testutils.serialization.types;
 
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.Random;
 
 import org.apache.flink.core.memory.DataInputView;
 import org.apache.flink.core.memory.DataOutputView;
 
-public class LongType implements SerializationTestType {
+public class ByteArrayType implements SerializationTestType {
 
-	private long value;
+	private static final int MAX_LEN = 512 * 15;
 
-	public LongType() {
-		this.value = 0;
+	private byte[] data;
+
+	public ByteArrayType() {
+		this.data = new byte[0];
 	}
 
-	private LongType(long value) {
-		this.value = value;
+	public ByteArrayType(byte[] data) {
+		this.data = data;
 	}
 
 	@Override
-	public LongType getRandom(Random rnd) {
-		return new LongType(rnd.nextLong());
+	public ByteArrayType getRandom(Random rnd) {
+		final int len = rnd.nextInt(MAX_LEN) + 1;
+		final byte[] data = new byte[len];
+		rnd.nextBytes(data);
+		return new ByteArrayType(data);
 	}
 
 	@Override
 	public int length() {
-		return 8;
+		return data.length + 4;
 	}
 
 	@Override
 	public void write(DataOutputView out) throws IOException {
-		out.writeLong(this.value);
+		out.writeInt(this.data.length);
+		out.write(this.data);
 	}
 
 	@Override
 	public void read(DataInputView in) throws IOException {
-		this.value = in.readLong();
+		final int len = in.readInt();
+		this.data = new byte[len];
+		in.readFully(this.data);
 	}
 
 	@Override
 	public int hashCode() {
-		return (int) (this.value ^ this.value >>> 32);
+		return Arrays.hashCode(this.data);
 	}
 
 	@Override
 	public boolean equals(Object obj) {
-		if (obj instanceof LongType) {
-			LongType other = (LongType) obj;
-			return this.value == other.value;
+		if (obj instanceof ByteArrayType) {
+			ByteArrayType other = (ByteArrayType) obj;
+			return Arrays.equals(this.data, other.data);
 		} else {
 			return false;
 		}
