+++ flink-core/src/test/java/org/apache/flink/testutils/serialization/types/ByteType.java
@@ -17,65 +17,56 @@
  */
 
 
-package org.apache.flink.runtime.io.network.api.serialization.types;
+package org.apache.flink.testutils.serialization.types;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.Random;
 
 import org.apache.flink.core.memory.DataInputView;
 import org.apache.flink.core.memory.DataOutputView;
 
-public class ByteArrayType implements SerializationTestType {
+public class ByteType implements SerializationTestType {
 
-	private static final int MAX_LEN = 512 * 15;
+	private byte value;
 
-	private byte[] data;
-
-	public ByteArrayType() {
-		this.data = new byte[0];
+	public ByteType() {
+		this.value = (byte) 0;
 	}
 
-	public ByteArrayType(byte[] data) {
-		this.data = data;
+	private ByteType(byte value) {
+		this.value = value;
 	}
 
 	@Override
-	public ByteArrayType getRandom(Random rnd) {
-		final int len = rnd.nextInt(MAX_LEN) + 1;
-		final byte[] data = new byte[len];
-		rnd.nextBytes(data);
-		return new ByteArrayType(data);
+	public ByteType getRandom(Random rnd) {
+		return new ByteType((byte) rnd.nextInt(256));
 	}
 
 	@Override
 	public int length() {
-		return data.length + 4;
+		return 1;
 	}
 
 	@Override
 	public void write(DataOutputView out) throws IOException {
-		out.writeInt(this.data.length);
-		out.write(this.data);
+		out.writeByte(this.value);
 	}
 
 	@Override
 	public void read(DataInputView in) throws IOException {
-		final int len = in.readInt();
-		this.data = new byte[len];
-		in.readFully(this.data);
+		this.value = in.readByte();
 	}
 
 	@Override
 	public int hashCode() {
-		return Arrays.hashCode(this.data);
+		return this.value;
 	}
 
 	@Override
 	public boolean equals(Object obj) {
-		if (obj instanceof ByteArrayType) {
-			ByteArrayType other = (ByteArrayType) obj;
-			return Arrays.equals(this.data, other.data);
+		if (obj instanceof ByteType) {
+			ByteType other = (ByteType) obj;
+			return this.value == other.value;
 		} else {
 			return false;
 		}
